requirements.txt
# Don't edit this file. It's generated automatically!
# If you want to update global dependencies, modify fixed-requirements.txt
# and then run 'make requirements' to update requirements.txt for all
# components.
# If you want to update depdencies for a single component, modify the
# in-requirements.txt for that component and then run 'make requirements' to
# update the component requirements.txt
RandomWords
amqp==5.0.6
apscheduler==3.7.0
argcomplete==1.12.2
bcrypt==3.2.0
chardet<3.1.0
cryptography==3.4.7
decorator==4.4.2
dnspython>=1.16.0,<2.0.0
eventlet==0.30.2
flex==6.14.1
git+https://github.com/StackStorm/logshipper.git@stackstorm_patched#egg=logshipper
git+https://github.com/StackStorm/orquesta.git@v1.5.0#egg=orquesta
git+https://github.com/StackStorm/st2-auth-backend-flat-file.git@master#egg=st2-auth-backend-flat-file
git+https://github.com/StackStorm/st2-auth-ldap.git@master#egg=st2-auth-ldap
git+https://github.com/StackStorm/st2-rbac-backend.git@master#egg=st2-rbac-backend
gitdb==4.0.2
gitpython==3.1.15
greenlet==1.0.0
gunicorn==20.1.0
importlib-metadata==3.10.1
jinja2==2.11.3
jsonpath-rw==1.4.0
jsonschema==2.6.0
kombu==5.0.2
lockfile==0.12.2
mock==4.0.3
mongoengine==0.23.0
networkx>=2.5.1,<2.6
nose
nose-parallel==0.4.0
nose-timer==1.0.1
orjson==3.5.2
oslo.config>=1.12.1,<1.13
oslo.utils<5.0,>=4.0.0
paramiko==2.7.2
passlib==1.7.4
prettytable==2.1.0
prompt-toolkit==1.0.15
psutil==5.8.0
pyinotify==0.9.6; platform_system=="Linux"
pymongo==3.11.3
pyrabbit
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger
python-statsd==2.1.0
pytz==2021.1
pywinrm==0.4.1
pyyaml==5.4.1
redis==3.5.3
rednose
requests[security]==2.25.1
retrying==1.3.3
routes==2.4.1
semver==2.13.0
simplejson
six==1.13.0
sseclient-py==1.7
stevedore==1.30.1
tooz==2.8.0
udatetime==0.0.16
unittest2
webob==1.8.7
webtest
zake==0.2.2
zstandard==0.15.2
